refactor: extract data-quality diagnostic contracts into focused module
Move OutlierSummaryResponse, DataQualityRequiredMissingPctResponse,
DataQualitySpeedCoverageResponse, DataQualityAccelSanityResponse,
DataQualityOutliersResponse, and DataQualityResponse from
history_analysis_contracts.py into shared/types/data_quality_contracts.py.

Update all consumers: summary_serialization/_summary.py,
adapters/http/models/history.py, and
tests/hygiene/test_api_model_boundary_parity.py.

Closes #1403
